pool-instanceと何の関係もないので別ファイルに切り出しておく
